Convert block ID  to byte

+++ src/game/blocks/BlockDefinition.java
@@ -25,7 +25,7 @@ public class BlockDefinition {
     private static final byte atlasSizeY = 32;
 
     //actual block object fields
-    public int     ID;
+    public byte     ID;
     public String  name;
     public boolean dropsItem;
     public float[] frontTexture;  //front
@@ -52,7 +52,7 @@ public class BlockDefinition {
     public String droppedItem;
 
     public BlockDefinition(
-            int ID,
+            byte ID,
             float stoneHardness,
             float dirtHardness,
             float woodHardness,
@@ -315,7 +315,7 @@ public class BlockDefinition {
 
 
         new BlockDefinition(
-                0,
+                (byte) 0,
                 -1f,
                 -1f,
                 -1f,
@@ -343,7 +343,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                1,
+                (byte) 1,
                 0f,
                 1f,
                 0f,
@@ -371,7 +371,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                2,
+                (byte) 2,
                 0,
                 2f,
                 0,
@@ -399,7 +399,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                3,
+                (byte) 3,
                 1,
                 0,
                 0,
@@ -427,7 +427,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                4,
+                (byte) 4,
                 1.5f,
                 0,
                 0,
@@ -455,7 +455,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                5,
+                (byte) 5,
                 -1,
                 -1,
                 -1,
@@ -492,7 +492,7 @@ public class BlockDefinition {
         };
 
         new BlockDefinition(
-                6,
+                (byte) 6,
                 0,
                 0,
                 2,
@@ -520,7 +520,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                7,
+                (byte) 7,
                 -1,
                 -1,
                 -1,
@@ -548,7 +548,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                8,
+                (byte) 8,
                 4,
                 0,
                 0,
@@ -576,7 +576,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                9,
+                (byte) 9,
                 6,
                 0,
                 0,
@@ -604,7 +604,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                10,
+                (byte) 10,
                 8,
                 0,
                 0,
@@ -632,7 +632,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                11,
+                (byte) 11,
                 10,
                 0,
                 0,
@@ -660,7 +660,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                12,
+                (byte) 12,
                 12,
                 0,
                 0,
@@ -688,7 +688,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                13,
+                (byte) 13,
                 10,
                 0,
                 0,
@@ -716,7 +716,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                14,
+                (byte) 14,
                 14,
                 0,
                 0,
@@ -744,7 +744,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                15,
+                (byte) 15,
                 16,
                 0,
                 0,
@@ -772,7 +772,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                16,
+                (byte) 16,
                 2,
                 0,
                 0,
@@ -801,7 +801,7 @@ public class BlockDefinition {
 
 
         new BlockDefinition(
-                17,
+                (byte) 17,
                 0,
                 0,
                 1,
@@ -828,7 +828,7 @@ public class BlockDefinition {
                 null
         );
         new BlockDefinition(
-                18,
+                (byte) 18,
                 0,
                 0,
                 1,
@@ -855,7 +855,7 @@ public class BlockDefinition {
                 null
         );
         new BlockDefinition(
-                19,
+                (byte)19,
                 0,
                 0,
                 1,
@@ -893,7 +893,7 @@ public class BlockDefinition {
             }
         };
         new BlockDefinition(
-                20,
+                (byte) 20,
                 0,
                 1,
                 0,
@@ -921,7 +921,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                21,
+                (byte) 21,
                 0,
                 0,
                 1,
@@ -967,7 +967,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                22,
+                (byte) 22,
                 0,
                 0,
                 1,
@@ -1014,7 +1014,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                23,
+                (byte) 23,
                 0,
                 0,
                 1,
@@ -1061,7 +1061,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                24,
+                (byte) 24,
                 0,
                 0,
                 1,
@@ -1108,7 +1108,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                25,
+                (byte) 25,
                 0,
                 0,
                 3,
@@ -1136,7 +1136,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                26,
+                (byte) 26,
                 0,
                 0,
                 0,
@@ -1164,7 +1164,7 @@ public class BlockDefinition {
         );
 
         new BlockDefinition(
-                27,
+                (byte) 27,
                 0,
                 0,
                 2,
@@ -1200,7 +1200,7 @@ public class BlockDefinition {
         };
 
         new BlockDefinition(
-                28,
+                (byte) 28,
                 0,
                 0,
                 2,
@@ -1241,7 +1241,7 @@ public class BlockDefinition {
         };
 
         new BlockDefinition(
-                29,
+                (byte) 29,
                 0,
                 0,
                 0,
